# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FitConnect API",
        "description": "Backend API for FitConnect — a fitness platform connecting trainers and clients. Provides authentication, client management, exercise/program libraries, session scheduling, real-time workout tracking, chat, notifications, packages, transactions, and progress metrics.",
        "contact": {
            "email": "support@fitconnect.app"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/package-templates": {
            "get": {
                "tags": [
                    "Packages"
                ],
                "summary": "List package templates",
                "operationId": "a943602ae66956b965eba250683c4e27",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of package templates",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PackageTemplate"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginatedResponse/properties/meta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Packages"
                ],
                "summary": "Create a new package template",
                "operationId": "04183f77e06e25dfb894f1d4f994d413",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "kind",
                                    "sessions_count",
                                    "validity_days",
                                    "price",
                                    "currency"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "kind": {
                                        "type": "string",
                                        "enum": [
                                            "count_based",
                                            "time_based",
                                            "hybrid"
                                        ]
                                    },
                                    "sessions_count": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "validity_days": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "price": {
                                        "type": "number"
                                    },
                                    "currency": {
                                        "type": "string"
                                    },
                                    "auto_renew_default": {
                                        "type": "boolean",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Package template created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/PackageTemplate"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/package-templates/{template}": {
            "get": {
                "tags": [
                    "Packages"
                ],
                "summary": "Get a package template",
                "operationId": "b43ad4ee1e9dec8db132ba10e28e0cfb",
                "parameters": [
                    {
                        "name": "template",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Package template retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/PackageTemplate"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Packages"
                ],
                "summary": "Update a package template",
                "operationId": "4399910e3bd736252d3bac35ddc7191e",
                "parameters": [
                    {
                        "name": "template",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "kind": {
                                        "type": "string",
                                        "nullable": true,
                                        "enum": [
                                            "count_based",
                                            "time_based",
                                            "hybrid"
                                        ]
                                    },
                                    "sessions_count": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "validity_days": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "price": {
                                        "type": "number",
                                        "nullable": true
                                    },
                                    "currency": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "auto_renew_default": {
                                        "type": "boolean",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Package template updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/PackageTemplate"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/package-templates/{template}/archive": {
            "post": {
                "tags": [
                    "Packages"
                ],
                "summary": "Archive a package template",
                "operationId": "e8f3bad8f52cf55a7bacb7eec41b1807",
                "parameters": [
                    {
                        "name": "template",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Package template archived",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/PackageTemplate"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PackageTemplate": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "trainer_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "kind": {
                        "type": "string",
                        "enum": [
                            "count_based",
                            "time_based",
                            "hybrid"
                        ]
                    },
                    "sessions_count": {
                        "type": "integer"
                    },
                    "validity_days": {
                        "type": "integer"
                    },
                    "price": {
                        "type": "number"
                    },
                    "currency": {
                        "type": "string"
                    },
                    "auto_renew_default": {
                        "type": "boolean"
                    },
                    "archived_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Sanctum access token",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}